Set the default working directory to be /var/lib/gitea
The default `app.ini` settings use a number of paths that are relative
to the working directory. Most of these paths are for backend data that
should not be located under the `install_dir`.

In addition to the `install_dir` being the wrong place for most of this
data, the gitea user does not have permissions to create directories
under the `install_dir`. One example of where this causes problems is
when trying to upload files via the web interface (which defaults to
using `TEMP_PATH = data/tmp/uploads`).

Setting the default working directory to `/var/lib/gitea` should resolve
most of these problems.