Don't automatically unset gitea.binary_hash if default binary is changed
The formula includes a default known working Gitea binary and associated
hash to verify the binary. If the user whats to use a different binary
then they need to either provide their own hash or explicitly disable
checking the binary hash by setting the `gitea:lookup:binary_hash`
pillar value to `None`.

This is to ensure that users are knowingly disabling the hash
verification, instead of the formula doing it for them without them
necessarily having any knowledge that no check are being done.