Set the GITEA_WORK_DIR in the gitea.service file
While we would assume that relative paths in the `app.ini` file would be
relative to the current working directory of the process, as of v1.4.0
this is not the case.

The relative paths are relative to the Gitea binary, unless the
`GITEA_WORK_DIR` environment variable is set, in which case the paths
are relative the `GITEA_WORK_DIR` value.

The should resolve the problems that were original mean to be fixed in
7cfc8c7076