Manage oauth2 settings
Gitea added support for oauth2 in v1.8.0 and it requires a JWT_SECRET
value to be set, even if oauth2 is disabled. If a JWT_SECRET value isn't
set then Gitea will attempt to write a value to the config files, which
will fail because the service doesn't have permission to do this. We
automatically generate a JWT_SECRET value if one isn't provided to avoid
this problem.